Split core assignment tests
